+++ CHANGELOG.md
@@ -5,11 +5,12 @@
 - Add method render_nothing to ruby and crystal functions, to make it clear how to send an empty response body
 - Empty functions will render an empty body
 - Ensure Watchdog listeners stop when exiting
-- Fix bug preventing create-local-dir from creating a folder when `create-workspace` is used
+- Fix bug preventing `--create-local-dir` from creating a folder when `create-workspace` is used
 - Change message confirmation when destroying a workspace
 - Update YAML template comments with mention to `test_command`
 - FaaStRuby Local won't start unless it can find `project.yml` in current directory.
 - Fix command `faastruby deploy`
+- Fix bug when response headers are empty
 
 ## 0.5.9 - Mar 11 2019
 - Add Gemfile support to Watchdog. When a `Gemfile` is added to a function, Watchdog will initialize it if the file is empty and run `bundle install` every time you modify it.